move shared models to a separate file
