+++ agenta-backend/agenta_backend/routers/configs_router.py
@@ -15,7 +15,7 @@ from agenta_backend.services import (
 )
 
 if isCloudEE():
-    from agenta_backend.commons.models.db_models import Permission
+    from agenta_backend.commons.models.shared_models import Permission
     from agenta_backend.commons.utils.permissions import check_action_access
