+++ agenta-backend/agenta_backend/routers/evaluators_router.py
@@ -15,7 +15,7 @@ from agenta_backend.models.api.evaluation_model import (
 )
 
 if isCloudEE():
-    from agenta_backend.commons.models.db_models import Permission
+    from agenta_backend.commons.models.shared_models import Permission
     from agenta_backend.commons.utils.permissions import check_action_access
 
 router = APIRouter()
