+++ agenta-backend/agenta_backend/routers/human_evaluation_router.py
@@ -27,7 +27,7 @@ from agenta_backend.services.evaluation_service import (
 )
 
 if isCloudEE():
-    from agenta_backend.commons.models.db_models import (
+    from agenta_backend.commons.models.shared_models import (
         Permission,
     )  # noqa pylint: disable-all
     from agenta_backend.commons.utils.permissions import (
